requirements.txt
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
